+++ BusinessLogic/Contracts/IIncomeService.cs
@@ -6,7 +6,7 @@ namespace BusinessLogic.Contracts
 {
     public interface IIncomeService : IFormService<IIncomeRepository, Income>
     {
-        IncomeViewModel GetByUserId(int userId, bool isOtherParent = false);
+        IncomeViewModel GetByUserId(long userId, bool isOtherParent = false);
         Income AddOrUpdate(IncomeViewModel model);
     }
 }
