+++ BusinessLogic/Contracts/IPreexistingSupportService.cs
@@ -7,7 +7,7 @@ namespace BusinessLogic.Contracts
 {
     public interface IPreexistingSupportService : IFormService<IPreexistingSupportRepository,PreexistingSupport>
     {
-        List<PreexistingSupport> GetByUserId(int userId, bool isOtherParent = false);
+        List<PreexistingSupport> GetByUserId(long userId, bool isOtherParent = false);
         PreexistingSupport AddOrUpdate(PreexistingSupportViewModel model);
     }
 }
